fix(backoffice): isolate marketplace client rendering
Render the Marketplace installation workflow lazily inside ClientOnly
so server rendering does not evaluate its client-only module. Move
generated UI selection into a presentation module to keep component
exports compatible with Fast Refresh.
